// src/app/Core/interfaces/user.model.ts
export interface User {
  avgCustomerRate: number | null;
  birthdate: string;
  createdAt: string;
  customerRateCount: number;
  email: string;
  fullName: string;
  genderId: number;
  id: number;
  modifiedAt: string | null;
  nationalityId: number;
  phoneNumber: string;
  picture: string | null;
}